Word-Clone/requirements.txt
docx2txt~=0.8
PyQt5~=5.15.9
pycryptodome~=3.15.0
google-auth-oauthlib~=1.2.0
google-auth-httplib2~=0.2.0
google-api-python-client~=2.121.0